IN PROGRESS:Fixed some PR comments. Moved some functions from routes to controllers/admin/changeTrackHoursVolunteer.
